dom/tests/mochitest/fetch/test_headers_sw_reroute.html



<!DOCTYPE HTML>
<html>
<head>
  <title>Test Fetch Headers - Basic</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<script type="text/javascript" src="sw_reroute.js"> </script>
<script class="testbody" type="text/javascript">
testScript("test_headers_common.js");
</script>
</body>
</html>
